Organize code into multiple files
